map: refactor
